French version / Correction de bug : -Ajoute TIF en plus de TIFF dans la liste des extensions supportées -Ajoute une icône dans la timeline pour indiquer les slides comprenant au moins un bloque avec transformation -Améliore encore la lecture des vidéos -Remet les boites de dialogue systèmes pour l'ouverture des fichiers
English version / Bug fix :
-Add TIF extension in the supported extension list
-Add an icon to the timeline to symbolise when at least one block of the slide have a transformation
-Improve again video reading
-Use system dialog box for opening files

git-svn-id: file:///media/werkstatt/064f12d4-5f27-4839-9537-9ffb7939d236/ffDiaporama-Entwicklung-2023/ffdiaporama-svn/trunk@278 f803d417-eb43-4c80-8848-1a7435ce69a7

+++ ffDiaporama/src/ffDiaporama/DlgImageCorrection.cpp
@@ -52,6 +52,11 @@ DlgImageCorrection::DlgImageCorrection(cCompositionObject *TheCurrentTextItem,in
      } else if (CurrentBrush->Video) {
         InitialFilteredString=CurrentBrush->Video->BrushFileTransform.FilterToString();
         CachedImage=CurrentBrush->Video->ImageAt(true,VideoPosition,QTime(0,0,0,0).msecsTo(CurrentBrush->Video->StartPos),NULL,1,false,NULL,false);
+        if (CachedImage->format()!=QImage::Format_ARGB32_Premultiplied) {
+            QImage *NewCachedImage=new QImage(CachedImage->convertToFormat(QImage::Format_ARGB32_Premultiplied));
+            delete CachedImage;
+            CachedImage=NewCachedImage;
+        }
         GlobalMainWindow->ApplicationConfig->StyleImageFramingCollection.SetImageGeometryFilter(GlobalMainWindow->Diaporama->ImageGeometry,CurrentBrush->Video->ObjectGeometry);
     }
 
@@ -366,7 +371,7 @@ void DlgImageCorrection::accept() {
     QString CachedFile=CurrentBrush->Image?CurrentBrush->Image->FileName:CurrentBrush->Video->FileName;
     CachedFile=CachedFile.replace("."+QFileInfo(CachedFile).suffix(),"_ffd.jpg");
     QString NewFilter=CurrentBrush->Image?CurrentBrush->Image->BrushFileTransform.FilterToString():CurrentBrush->Video->BrushFileTransform.FilterToString();
-    if ((NewFilter!="")&&((InitialFilteredString!=NewFilter)||((GlobalMainWindow->ApplicationConfig->AllowCachedTransfoImages)&&(!QFileInfo(CachedFile).exists())))) {
+    if ((NewFilter!="")&&(CurrentBrush->Image)&&((InitialFilteredString!=NewFilter)||((GlobalMainWindow->ApplicationConfig->AllowCachedTransfoImages)&&(!QFileInfo(CachedFile).exists())))) {
         if (QFileInfo(CachedFile).exists()) QFile::remove(CachedFile);
         if (GlobalMainWindow->ApplicationConfig->AllowCachedTransfoImages) {
             QApplication::setOverrideCursor(QCursor(Qt::WaitCursor));
@@ -374,15 +379,9 @@ void DlgImageCorrection::accept() {
             screen.showMessage(QApplication::translate("DlgImageCorrection","Creating cached filtered file ..."),Qt::AlignHCenter|Qt::AlignBottom);
             screen.show();
             QApplication::processEvents();
-            cLuLoImageCacheObject *ImageObject=GlobalMainWindow->ApplicationConfig->ImagesCache.FindObject(
-                        CurrentBrush->Image?CurrentBrush->Image->FileName:CurrentBrush->Video->FileName,
-                        CurrentBrush->Image?CurrentBrush->Image->ModifDateTime:CurrentBrush->Video->ModifDateTime,
-                        CurrentBrush->Image?CurrentBrush->Image->ImageOrientation:CurrentBrush->Video->ImageOrientation,
-                        NULL,true,true);
-
+            cLuLoImageCacheObject *ImageObject=GlobalMainWindow->ApplicationConfig->ImagesCache.FindObject(CurrentBrush->Image->FileName,CurrentBrush->Image->ModifDateTime,CurrentBrush->Image->ImageOrientation,NULL,true,true);
             QImage *UnfilteredImage=new QImage(ImageObject->ValidateCacheRenderImage()->copy());
-            if (CurrentBrush->Image) CurrentBrush->Image->BrushFileTransform.ApplyFilter(UnfilteredImage);
-                else                 CurrentBrush->Video->BrushFileTransform.ApplyFilter(UnfilteredImage);
+            CurrentBrush->Image->BrushFileTransform.ApplyFilter(UnfilteredImage);
             UnfilteredImage->save(CachedFile,"jpg",100);
             delete UnfilteredImage;
             screen.hide();
@@ -662,6 +661,11 @@ void DlgImageCorrection::ChangeBrushDiskFile() {
         delete CachedImage;
         CachedImage=NewCachedImage;
         UndoReloadImage=true;
+    }
+    if (CachedImage->format()!=QImage::Format_ARGB32_Premultiplied) {
+        QImage *NewCachedImage=new QImage(CachedImage->convertToFormat(QImage::Format_ARGB32_Premultiplied));
+        delete CachedImage;
+        CachedImage=NewCachedImage;
     }
     RefreshControls();
 }
